[FIX] Edgecase with event queue never getting cleaned
- What happened:
-- due to memory issues an application using this library was not able
to spawn a new threads for a short time. During this time the plugin
tried to spawn the worker again. Before even doing so it sets
"working=true" but "working=false" is only ever set from within the
thread, so if the thread fails to start it is not endlessly stuck in
"working==true" and will never spawn another worker thread that cleans
the even queue
- The fix now ensures that working is only set to true within the thread
as well assuming it has not happened from some other thread before in
which case the thread returns right away